% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_HNSCatlas.R
\name{load_HNSCatlas}
\alias{load_HNSCatlas}
\title{Download and Load HNSCatlas Data}
\usage{
load_HNSCatlas(destpath = tempdir())
}
\arguments{
\item{destpath}{Directory or full file path where the dataset will be stored.
If a directory is provided, the dataset will be saved as "HNSCatlas.rds".
Default is the temporary directory.}
}
\value{
A Seurat object containing the HNSCatlas dataset.
}
\description{
This function downloads the HNSCatlas dataset if it is not already present
and loads it into the R environment as a Seurat object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
HNSCatlas <- load_HNSCatlas()  # Uses temp directory
HNSCatlas <- load_HNSCatlas("/omics/odcf/analysis/OE0509_projects/hnscc/single_cell")
\dontshow{\}) # examplesIf}
}
